# public/swagger.json
{
	"components": {
...
		"schemas": {
			"EventParamsType": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double"
					},
					"eventName": {
						"type": "string"
					},
					"profile": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"txId": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"eventName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorVo": {
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"messageCode": {
						"type": "string"
					},
					"events": {
						"items": {
							"$ref": "#/components/schemas/EventParamsType"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"message",
					"messageCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommonResponse": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"CommonResponse_any_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"App": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdId": {
						"type": "number",
						"format": "double"
					},
					"siteUrl": {
						"type": "string",
						"nullable": true
					},
					"detailDescription": {
						"type": "string",
						"nullable": true
					},
					"bannerUrl": {
						"type": "string",
						"nullable": true
					},
					"promotionalText": {
						"type": "string",
						"nullable": true
					},
					"cardUrl": {
						"type": "string",
						"nullable": true
					},
					"blockChainId": {
						"type": "string"
					},
					"iconUrl": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"appId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"updatedId",
					"createdId",
					"siteUrl",
					"detailDescription",
					"bannerUrl",
					"promotionalText",
					"cardUrl",
					"blockChainId",
					"iconUrl",
					"name",
					"appId"
				],
				"type": "object",
				"description": "Model App\nThis model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments"
			},
			"CommonResponse_App_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/App"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"Platform": {
				"type": "string",
				"enum": [
					"iOS",
					"Android",
					"Steam",
					"Window",
					"macOS"
				]
			},
			"PlatformUrl": {
				"properties": {
					"platform": {
						"$ref": "#/components/schemas/Platform"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"platform",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SupportChainId": {
				"enum": [
					137,
					80001,
					56,
					97
				],
				"type": "number"
			},
...
			"LanguageOptions": {
				"type": "string",
				"enum": [
					"ko",
					"en",
					"cn",
					"jp"
				]
			},
			"ApplyReleaseRequest": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double",
						"description": "앱 id",
						"example": 1
					},
					"promotionalText": {
						"type": "string",
						"description": "소개글",
						"example": "소개글"
					},
					"bannerUrl": {
						"type": "string",
						"description": "배너 이미지 url",
						"example": "http://example.image/banner"
					},
					"cardUrl": {
						"type": "string",
						"description": "카드 이미지 url",
						"example": "http://example.image/card"
					},
					"screenUrls": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "스크린샷 이미지 url 목록",
						"example": [
							"http://example.image/screenshot1",
							"http://example.image/screenshot2"
						]
					},
					"detailDescription": {
						"type": "string",
						"description": "상세 설명",
						"example": "상세설명"
					},
					"siteUrl": {
						"type": "string",
						"description": "홈페이지 주소",
						"example": "http://example.homepage.com"
					},
					"platformUrls": {
						"items": {
							"$ref": "#/components/schemas/PlatformUrl"
						},
						"type": "array",
						"description": "플랫폼 링크\r\n플랫폼 종류 : \"iOS\" , \"Android\" , \"Steam\" , \"Window\", \"macOS\""
					},
					"selectedLanguages": {
						"items": {
							"$ref": "#/components/schemas/LanguageOptions"
						},
						"type": "array",
						"description": "지원 언어 \"ko\" , \"en\" , \"cn\" , \"jp\""
					}
				},
				"required": [
					"appId",
					"promotionalText",
					"bannerUrl",
					"cardUrl",
					"screenUrls",
					"detailDescription",
					"siteUrl",
					"platformUrls",
					"selectedLanguages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FindManyReleaseHistoryResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "출시 히스토리 id"
					},
					"appId": {
						"type": "number",
						"format": "double",
						"description": "앱 id"
					},
					"reviewId": {
						"type": "string",
						"description": "출시 심사 id"
					},
					"statusCd": {
						"type": "string",
						"description": "심사상태코드. 코드그룹(APRV_STAT), REQUEST: 심사요청, IN_REVIEW: 심사중, REJECTED: 반려, RELEASED: 출시, CANCELED: 취소됨"
					},
					"createdId": {
						"type": "number",
						"format": "double",
						"description": "작성자 id"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "작성일"
					}
				},
				"required": [
					"id",
					"appId",
					"reviewId",
					"statusCd",
					"createdId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_FindManyReleaseHistoryResponse-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/FindManyReleaseHistoryResponse"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"TokenTransferRequest": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double"
					},
					"walletId": {
						"type": "number",
						"format": "double"
					},
					"symbol": {
						"type": "string"
					},
					"toAddress": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"chainId": {
						"$ref": "#/components/schemas/SupportChainId"
					}
				},
				"required": [
					"walletId",
					"symbol",
					"toAddress",
					"amount",
					"chainId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransferNativeResponse": {
				"properties": {
					"fromAddress": {
						"type": "string",
						"description": "sender's wallet address",
						"example": "0xBcb0a748ea81B17274De6714ea60BD56E76E11cb"
					},
					"toAddress": {
						"type": "string",
						"description": "recipient wallet address",
						"example": "0xBcb0a748ea81B17274De6714ea60BD56E76E11cb"
					},
					"amount": {
						"type": "number",
						"format": "double",
						"description": "amount of native token",
						"example": 1
					},
					"txHash": {
						"type": "string",
						"description": "transaction hash",
						"example": "0x426eb17979ade27411a8209ff65305c14e892268bc66a074017ddd4a9c3f7d7e"
					},
					"rowData": {
						"description": "rawdata from rpc result including receipt, logs, event"
					}
				},
				"required": [
					"amount",
					"txHash",
					"rowData"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransferTokenResponse": {
				"$ref": "#/components/schemas/TransferNativeResponse"
			},
			"CommonResponse_TransferTokenResponse_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/TransferTokenResponse"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"TransferTokenRequest": {
				"properties": {
					"toAddress": {
						"type": "string",
						"description": "recipient wallet address",
						"example": "0xBcb0a748ea81B17274De6714ea60BD56E76E11cb"
					},
					"amount": {
						"type": "number",
						"format": "double",
						"description": "amount of native token",
						"example": 1
					},
					"appId": {
						"type": "number",
						"format": "double"
					},
					"tokenContractAddress": {
						"type": "string",
						"description": "Include the smart contract address of the token you want to transfer",
						"example": "0x507f60A808C4e69E4Af0395F9265C7F89CE03D04"
					},
					"privateKey": {
						"type": "string",
						"description": "private key",
						"example": ""
					}
				},
				"required": [
					"toAddress",
					"amount",
					"tokenContractAddress",
					"privateKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FindWalletByAppIdQueryResponse": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double",
						"description": "앱 id"
					},
					"walletId": {
						"type": "string",
						"description": "콘솔 지갑 id"
					},
					"address": {
						"type": "string",
						"description": "콘솔 지갑 주소"
					},
					"encPrivateKey": {
						"type": "string",
						"description": "암호화 프라이빗키"
					},
					"autoSignYn": {
						"type": "string",
						"enum": [
							"Y",
							"N"
						],
						"description": "자동 서명 여부"
					},
					"tokenCount": {
						"type": "number",
						"format": "double",
						"description": "보유 토큰 수량"
					},
					"PMG_balance": {
						"type": "string",
						"description": "보유 PMG(기준 재화 토큰) 토큰 수량"
					},
					"memberAccessYn": {
						"type": "string",
						"enum": [
							"Y",
							"N"
						],
						"description": "지갑 설정 멤버 접근 허용 여부"
					},
					"createdId": {
						"type": "number",
						"format": "double",
						"description": "앱 생성자 id"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "콘솔 지갑 생성일"
					}
				},
				"required": [
					"appId",
					"walletId",
					"address",
					"encPrivateKey",
					"autoSignYn",
					"tokenCount",
					"PMG_balance",
					"memberAccessYn",
					"createdId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_FindWalletByAppIdQueryResponse_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/FindWalletByAppIdQueryResponse"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"WalletTransactionDataType": {
				"properties": {
					"txId": {
						"type": "string"
					},
					"nonce": {
						"type": "string"
					},
					"blockNumber": {
						"type": "string"
					},
					"receiptStatus": {
						"type": "string"
					},
					"fromAddress": {
						"type": "string"
					},
					"toAddress": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"gas": {
						"type": "string",
						"nullable": true
					},
					"gasPrice": {
						"type": "string"
					},
					"blockTimestamp": {
						"type": "string"
					}
				},
				"required": [
					"txId",
					"nonce",
					"blockNumber",
					"receiptStatus",
					"fromAddress",
					"toAddress",
					"value",
					"gasPrice",
					"blockTimestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WalletTransactionsResponseType": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"page_size": {
						"type": "number",
						"format": "double"
					},
					"cursor": {
						"type": "string",
						"nullable": true
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/WalletTransactionDataType"
						},
						"type": "array"
					},
					"hasNext": {
						"type": "boolean"
					}
				},
				"required": [
					"page",
					"page_size",
					"cursor",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_WalletTransactionsResponseType_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/WalletTransactionsResponseType"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"devAuth": {
				"type": "http",
				"scheme": "basic"
			},
			"basic": {
				"type": "http",
				"scheme": "basic"
			},
			"token": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"cookie": {
				"type": "apiKey",
				"name": "accessToken",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "platpome-console-backend",
		"version": "0.0.1",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/release": {
			"post": {
				"operationId": "ApplyRelease",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_App_"
								}
							}
						}
					}
				},
				"description": "앱 출시 심사 요청 API",
				"tags": [
					"[출시] Release"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApplyReleaseRequest"
							}
						}
					}
				}
			}
		},
		"/release/{appId}": {
			"put": {
				"operationId": "CancelApplyRelease",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse"
								}
							}
						}
					}
				},
				"description": "출시 심사 신청 취소 API",
				"tags": [
					"[출시] Release"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"get": {
				"operationId": "GetReleaseHistories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_FindManyReleaseHistoryResponse-Array_"
								}
							}
						}
					}
				},
				"description": "앱 심사 내역 목록 조회 API",
				"tags": [
					"[출시] Release"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/release/approve/{reviewId}": {
			"post": {
				"operationId": "ApproveRelease",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_App_"
								}
							}
						}
					}
				},
				"description": "앱 출시 심사 승인 API",
				"tags": [
					"[출시] Release"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "reviewId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tokens": {
			"post": {
				"operationId": "TransferToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse"
								}
							}
						}
					}
				},
				"description": "ERC20 토큰 전송 API",
				"tags": [
					"[토큰] Token"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TokenTransferRequest"
							}
						}
					}
				}
			}
		},
		"/wallets/{appId}/member-access/{walletId}": {
			"put": {
				"operationId": "ToggleMemberAccess",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_any_"
								}
							}
						}
					}
				},
				"description": "지갑 설정 및 기능 마스터 외 멤버 접근 허용/비허용 처리 API",
				"tags": [
					"[Wallet] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/wallets/{appId}/auto-sign/{walletId}": {
			"put": {
				"operationId": "ToggleAutoSign",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"error": {
											"$ref": "#/components/schemas/ErrorVo"
										},
										"data": {}
									},
									"required": [
										"total",
										"error",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "지갑 자동서명 허용/비허용 처리 API",
				"tags": [
					"[Wallet] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/wallets/{chainId}/wallets/transfer-token": {
			"post": {
				"operationId": "TransferToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_TransferTokenResponse_"
								}
							}
						}
					}
				},
				"description": "ERC20 토큰 전송 샘플",
				"tags": [
					"[Wallet] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "- 80001:mumbai testnet, 137:polygon mainnet",
						"in": "path",
						"name": "chainId",
						"required": true,
						"schema": {
							"default": 80001,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- true: include raw data, false: exclude raw data",
						"in": "query",
						"name": "includeRawData",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TransferTokenRequest"
							}
						}
					}
				}
			}
		},
		"/wallets/{chainId}/wallets/{appId}/detail": {
			"get": {
				"operationId": "FindWalletDetailByAppId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_FindWalletByAppIdQueryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"[Wallet] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "chainId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SupportChainId"
						}
					}
				]
			}
		},
		"/wallets/{chainId}/wallets/{appId}/address": {
			"get": {
				"operationId": "FindWalletByAppId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_FindWalletByAppIdQueryResponse_"
								}
							}
						}
					}
				},
				"description": "지갑 정보 조회 API",
				"tags": [
					"[Wallet] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "chainId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SupportChainId"
						}
					}
				]
			}
		},
		"/wallets/{chainId}/wallets/{walletAddress}/transactions": {
			"get": {
				"operationId": "GetTransactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_WalletTransactionsResponseType_"
								}
							}
						}
					}
				},
				"description": "지갑 transaction 내역 조회\r\ndocs: https://docs.moralis.io/web3-data-api/evm/reference/get-decoded-wallet-transaction?address=0xd8da6bf26964af9d7eed9e03e53415d37aa96045&chain=eth&include=internal_transactions\r\nTODO: fromDate, toDate, include 옵션 추가",
				"tags": [
					"[Wallet] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "- 97:BSC testnet, 56:BSC mainnet, 80001:mumbai testnet, 137:polygon mainnet",
						"in": "path",
						"name": "chainId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SupportChainId"
						}
					},
					{
						"description": "- wallet address",
						"in": "path",
						"name": "walletAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The desired page size of the result.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						},
						"example": 10
					},
					{
						"in": "query",
						"name": "fromDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "toDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The cursor returned in the previous response (used for getting the next page).",
						"in": "query",
						"name": "pageCursor",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJjdXN0b21QYXJhbXMiOnsidG9rZW5BZGRyZXNzIjoiMHg2YTJmNjlmNTc5YmIzMmMyNDBmNDZjNGZiNGQzMGM4YzhiY2NiYTQxIn0sImtleXMiOlsiMTY5MzI3NzkxNC45NSJdLCJ3aGVyZSI6eyJ0b2tlbl9hZGRyZXNzIjoiMHg2YTJmNjlmNTc5YmIzMmMyNDBmNDZjNGZiNGQzMGM4YzhiY2NiYTQxIn0sImxpbWl0IjoxMSwib2Zmc2V0IjowLCJvcmRlciI6W10sImRpc2FibGVfdG90YWwiOnRydWUsImV4Y2x1ZGVfc3BhbSI6ZmFsc2UsInRvdGFsIjpudWxsLCJwYWdlIjoxLCJ0YWlsT2Zmc2V0IjoxLCJpYXQiOjE2OTQwNjcyNzZ9.mv1I78rtooJ7_rP03CpQHuqj9J7ds1VucZVu1yaGYZQ"
					}
				]
			}
		},
		"/wallets/{chainId}/{walletAddress}/balances": {
			"get": {
				"operationId": "GetWalletNFTTokens",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"$ref": "#/components/schemas/ErrorVo"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {}
									},
									"required": [
										"error",
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "지갑이 보유한 NFT 목록 조회 API",
				"tags": [
					"[Wallet] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chainId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SupportChainId"
						}
					},
					{
						"in": "path",
						"name": "walletAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}